// src/types/report.ts
export interface itemReport {
  ID: any;
  ProductKey: any;
  IDUser: number;
  IDXeOTo: any;
  BienSoXe: string;
  NoiDungSuaChua: string;
  SoLuong: number;
  DonGia: number;
  ThanhTien: number;
  LaiXe: string;
  IDLaiXe: any;
  NgaySuaCal: string;
  NgaySua: string;
  NgayHoanThanhCal: string;
  NgayHoanThanh: string;
  NgayDoDauCal: any;
  NgayDoDau: any;
  GARAGE: string;
  GhiChu: any;
}